# docs/swagger.yaml
definitions:
...
  github_com_Vovarama1992_emelya-go_internal_money_tariff_model.Tariff:
    properties:
      block_until:
        type: string
      created_at:
        type: string
      daily_reward:
        type: number
      id:
        type: integer
      name:
        type: string
    type: object
...
  internal_money_reward_delivery.AdminCreateReferralRewardRequest:
    properties:
      amount:
        type: number
      user_id:
        type: integer
    required:
    - amount
    - user_id
    type: object
  internal_money_tariff_delivery.CreateTariffRequest:
    properties:
      block_until:
        description: ISO8601 строка или nil
        type: string
      daily_reward:
        type: number
      name:
        type: string
    required:
    - name
    type: object
  internal_money_tariff_delivery.UpdateTariffRequest:
    properties:
      block_until:
        type: string
      daily_reward:
        type: number
      id:
        type: integer
      name:
        type: string
    required:
    - id
    - name
    type: object
...
info:
  contact: {}
paths:
...
  /api/admin/reward/by-user:
    get:
      parameters:
      - description: ID пользователя
        in: query
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items: {}
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 'Админ: получить все награды пользователя'
      tags:
      - admin-reward
  /api/admin/reward/referral-income:
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные о вознаграждении
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/internal_money_reward_delivery.AdminCreateReferralRewardRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 'Админ: начислить доход от реферала'
      tags:
      - admin-reward
  /api/admin/reward/total-available:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: number
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить общую сумму доступных к выводу средств (агрегация по всем
        пользователям)
      tags:
      - admin-reward
  /api/admin/tariffs:
    delete:
      parameters:
      - description: ID тарифа
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Удалить тариф
      tags:
      - tariff
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_Vovarama1992_emelya-go_internal_money_tariff_model.Tariff'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить все тарифы
      tags:
      - tariff
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные тарифа
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/internal_money_tariff_delivery.CreateTariffRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Создать тариф
      tags:
      - tariff
    put:
      consumes:
      - application/json
      parameters:
      - description: Обновляемые данные тарифа
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/internal_money_tariff_delivery.UpdateTariffRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Обновить тариф
      tags:
      - tariff
...
  /api/reward/my:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items: {}
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 'Юзер: получить свои вознаграждения'
      tags:
      - reward
...
swagger: "2.0"